requirements.txt
Flask
pymupdf
openai>=1.0.0
Pillow
